requirements.txt
colorama==0.4.3
discord.py==1.7.3
discord_webhook==0.16.3
python-decouple==3.6
